Add support for LCD filters
Freetype is able to render glyphs at 3 times the width to exploit the
subpixel layout on LCD monitors and increase the apparent horizontal
resolution, at the expense of reduced color fidelity. This is achieved
by loading a glyph with the `FT_LOAD_TARGET_LCD` flag. Unfortunately,
this introduces severe color fringing. To remedy this, Freetype can
apply a 1D FIR filter along the X-axis to such bitmaps, reducing color
fringing considerably but with slightly blurrier results. The details
are better explained in [1].

This commit adds bindings for `FT_Library_SetLcdFilter` and the
`LcdFilter` enum in order to expose this functionality.

[1]: https://www.freetype.org/freetype2/docs/reference/ft2-lcd_filtering.html